@@ -1,6 +1,6 @@
 {
   "name": "@danielmoncada/angular-datetime-picker",
-  "version": "16.1.0",
+  "version": "17.0.0",
   "description": "Angular Date Time Picker",
   "keywords": [
     "Angular",
@@ -36,32 +36,32 @@
     "url": "https://github.com/danielmoncada/date-time-picker.git"
   },
   "dependencies": {
-    "@angular/animations": "^13.3.5",
-    "@angular/common": "^13.1.2",
-    "@angular/compiler": "^13.3.5",
-    "@angular/core": "^13.1.2",
-    "@angular/forms": "^13.3.5",
-    "@angular/platform-browser": "^13.3.5",
-    "@angular/platform-browser-dynamic": "^13.3.5",
-    "@angular/router": "^13.3.5",
+    "@angular/animations": "^17.0.4",
+    "@angular/common": "^17.0.4",
+    "@angular/compiler": "^17.0.4",
+    "@angular/core": "^17.0.4",
+    "@angular/forms": "^17.0.4",
+    "@angular/platform-browser": "^17.0.4",
+    "@angular/platform-browser-dynamic": "^17.0.4",
+    "@angular/router": "^17.0.4",
     "core-js": "^3.22.3",
-    "moment": "^2.29.3",
-    "moment-timezone": "^0.5.34",
-    "np": "^6.5.0",
-    "rxjs": "^6.6.7",
+    "moment": "^2.29.4",
+    "moment-timezone": "^0.5.43",
+    "np": "^8.0.4",
+    "rxjs": "^7.8.1",
     "tslib": "^2.4.0",
-    "uuid": "^8.3.2",
-    "zone.js": "~0.11.5"
+    "uuid": "^9.0.1",
+    "zone.js": "^0.14.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.3.4",
-    "@angular/cdk": "^13.3.5",
-    "@angular/cli": "^13.3.4",
-    "@angular/compiler-cli": "^13.3.5",
-    "@angular/language-service": "^13.3.5",
+    "@angular-devkit/build-angular": "^17.0.3",
+    "@angular/cdk": "^17.0.1",
+    "@angular/cli": "^17.0.3",
+    "@angular/compiler-cli": "^17.0.4",
+    "@angular/language-service": "^17.0.4",
     "@types/jasmine": "^3.10.6",
     "@types/jasminewd2": "^2.0.10",
-    "@types/node": "^12.20.50",
+    "@types/node": "^20.10.0",
     "codecov": "^3.8.3",
     "codelyzer": "^6.0.2",
     "jasmine-core": "~4.1.0",
@@ -71,16 +71,11 @@
     "karma-coverage": "^2.2.0",
     "karma-jasmine": "~5.0.0",
     "karma-jasmine-html-reporter": "^1.7.0",
-    "ng-packagr": "^13.3.1",
-    "node-sass": "^7.0.0",
+    "ng-packagr": "^17.0.2",
+    "node-sass": "^9.0.0",
     "protractor": "~7.0.0",
-    "ts-node": "^8.10.2",
+    "ts-node": "^10.9.1",
     "tslint": "^6.1.3",
-    "typescript": "~4.5.4"
-  },
-  "peerDependencies": {
-    "@angular/common": "^13.0.3 || ^14.0.0 || ^15.0.0 || ^16.0.0",
-    "@angular/core": "^13.0.3 || ^14.0.0 || ^15.0.0 || ^16.0.0",
-    "@angular/cdk": "^13.0.3 || ^14.0.0 || ^15.0.0 || ^16.0.0"
+    "typescript": "~5.2.2"
   }
 }
