@@ -1,6 +1,6 @@
 {
   "name": "@danielmoncada/angular-datetime-picker",
-  "version": "16.1.0",
+  "version": "17.0.0",
   "description": "Angular Date Time Picker",
   "keywords": [
     "Angular",
@@ -24,9 +24,9 @@
     "access": "public"
   },
   "peerDependencies": {
-    "@angular/common": "^13.0.3 || ^14.0.0 || ^15.0.0 || ^16.0.0",
-    "@angular/core": "^13.0.3 || ^14.0.0 || ^15.0.0 || ^16.0.0",
-    "@angular/cdk": "^13.0.3 || ^14.0.0 || ^15.0.0 || ^16.0.0"
+    "@angular/common": "^17.0.0",
+    "@angular/core": "^17.0.0",
+    "@angular/cdk": "^17.0.0"
   },
   "dependencies": {
     "tslib": "^2.3.1"
