@@ -236,13 +236,13 @@ export class OwlDateTimeContainerComponent<T>
     public handleContainerAnimationStart(event: AnimationEvent): void {
         const toState = event.toState;
         if (toState === 'enter') {
-            this.beforePickerOpened$.next();
+            this.beforePickerOpened$.next(null);
         }
     }
     public handleContainerAnimationDone(event: AnimationEvent): void {
         const toState = event.toState;
         if (toState === 'enter') {
-            this.pickerOpened$.next();
+            this.pickerOpened$.next(null);
         }
     }
 
