@@ -2,13 +2,12 @@
  * dialog-ref.class
  */
 import { AnimationEvent } from '@angular/animations';
-import { Location } from '@angular/common';
-import { GlobalPositionStrategy, OverlayRef } from '@angular/cdk/overlay';
 import { ESCAPE } from '@angular/cdk/keycodes';
-import { OwlDialogContainerComponent } from './dialog-container.component';
+import { GlobalPositionStrategy, OverlayRef } from '@angular/cdk/overlay';
+import { Location } from '@angular/common';
+import { Observable, Subject, Subscription, SubscriptionLike, filter, take } from 'rxjs';
 import { DialogPosition } from './dialog-config.class';
-import { Observable, Subject, Subscription, SubscriptionLike as ISubscription } from 'rxjs';
-import { filter, take } from 'rxjs/operators';
+import { OwlDialogContainerComponent } from './dialog-container.component';
 
 export class OwlDialogRef<T> {
 
@@ -23,7 +22,7 @@ export class OwlDialogRef<T> {
     private _afterClosed$ = new Subject<any>();
 
     /** Subscription to changes in the user's location. */
-    private locationChanged: ISubscription = Subscription.EMPTY;
+    private locationChanged: SubscriptionLike = Subscription.EMPTY;
 
     /**
      * The instance of component opened into modal
@@ -31,12 +30,13 @@ export class OwlDialogRef<T> {
     public componentInstance: T;
 
     /** Whether the user is allowed to close the dialog. */
-    public disableClose = this.container.config.disableClose;
+    public disableClose = true;
 
-    constructor( private overlayRef: OverlayRef,
-                 private container: OwlDialogContainerComponent,
-                 public readonly id: string,
-                 location?: Location ) {
+    constructor(private overlayRef: OverlayRef,
+        private container: OwlDialogContainerComponent,
+        public readonly id: string,
+        location?: Location) {
+        this.disableClose = this.container.config.disableClose;
 
         this.container.animationStateChanged
             .pipe(
@@ -44,7 +44,7 @@ export class OwlDialogRef<T> {
                 take(1)
             )
             .subscribe(() => {
-                this._beforeOpen$.next();
+                this._beforeOpen$.next(null);
                 this._beforeOpen$.complete();
             });
 
@@ -54,13 +54,13 @@ export class OwlDialogRef<T> {
                 take(1)
             )
             .subscribe(() => {
-                this._afterOpen$.next();
+                this._afterOpen$.next(null);
                 this._afterOpen$.complete();
             });
 
         this.container.animationStateChanged
             .pipe(
-                filter(( event: AnimationEvent ) => event.phaseName === 'done' && event.toState === 'exit'),
+                filter((event: AnimationEvent) => event.phaseName === 'done' && event.toState === 'exit'),
                 take(1)
             )
             .subscribe(() => {
@@ -84,12 +84,12 @@ export class OwlDialogRef<T> {
         }
     }
 
-    public close( dialogResult?: any ) {
+    public close(dialogResult?: any) {
         this.result = dialogResult;
 
         this.container.animationStateChanged
             .pipe(
-                filter(( event: AnimationEvent ) => event.phaseName === 'start'),
+                filter((event: AnimationEvent) => event.phaseName === 'start'),
                 take(1)
             )
             .subscribe(() => {
@@ -119,7 +119,7 @@ export class OwlDialogRef<T> {
      * Updates the dialog's position.
      * @param position New dialog position.
      */
-    public updatePosition( position?: DialogPosition ): this {
+    public updatePosition(position?: DialogPosition): this {
         const strategy = this.getPositionStrategy();
 
         if (position && (position.left || position.right)) {
@@ -144,7 +144,7 @@ export class OwlDialogRef<T> {
      * @param width New width of the dialog.
      * @param height New height of the dialog.
      */
-    updateSize( width: string = 'auto', height: string = 'auto' ): this {
+    updateSize(width: string = 'auto', height: string = 'auto'): this {
         this.getPositionStrategy().width(width).height(height);
         this.overlayRef.updatePosition();
         return this;
